Move time lanes model to new file
