+++ caqtus/gui/condetrol/timelanes_editor/__init__.py
@@ -1,12 +1,11 @@
 from ._delegate import TimeLaneDelegate
-from .digital_lane_model import DigitalTimeLaneModel
-from .analog_lane_model import AnalogTimeLaneModel
-from .model import TimeLanesModel, TimeLaneModel
 from ._time_lanes_editor import TimeLanesEditor
+from .analog_lane_model import AnalogTimeLaneModel
+from .digital_lane_model import DigitalTimeLaneModel
+from .model import TimeLaneModel
 
 __all__ = [
     "TimeLanesEditor",
-    "TimeLanesModel",
     "DigitalTimeLaneModel",
     "TimeLaneModel",
     "TimeLaneDelegate",
